Add guard in StorageExplorer onDeleteSelectedFiles

+++ studio/components/to-be-cleaned/Storage/StorageExplorer/StorageExplorer.js
@@ -234,6 +234,8 @@ const StorageExplorer = observer(({ bucket }) => {
   const onDeleteSelectedFiles = async () => {
     if (selectedItemsToDelete.length === 1) {
       const [itemToDelete] = selectedItemsToDelete
+      if (!itemToDelete) return
+
       switch (itemToDelete.type) {
         case STORAGE_ROW_TYPES.FOLDER:
           await deleteFolder(itemToDelete)
